% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{filter_variants}
\alias{filter_variants}
\title{Filter variants by standard quality statistics}
\usage{
filter_variants(
  x,
  min_maf = 0.05,
  max_maf = 0.5,
  max_missing = 0.25,
  max_het = 1,
  min_hwe_p = 0,
  quiet = FALSE
)
}
\arguments{
\item{x}{An `ld_data` object.}

\item{min_maf}{Minimum minor allele frequency.}

\item{max_maf}{Maximum minor allele frequency.}

\item{max_missing}{Maximum missing-call fraction.}

\item{max_het}{Maximum heterozygous-call fraction.}

\item{min_hwe_p}{Minimum chi-square Hardy-Weinberg P value; use zero to disable.}

\item{quiet}{Suppress the filtering summary.}
}
\value{
A filtered `ld_data` object.
}
\description{
Filter variants by standard quality statistics
}
