% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ld.R
\name{haplotype_frequencies}
\alias{haplotype_frequencies}
\title{Estimate haplotype frequencies}
\usage{
haplotype_frequencies(
  x,
  variants = NULL,
  min_frequency = 0.01,
  drop_missing = TRUE
)
}
\arguments{
\item{x}{An `ld_data` object.}

\item{variants}{Variant IDs or indices; all variants by default.}

\item{min_frequency}{Minimum reported frequency.}

\item{drop_missing}{Drop chromosomes/samples with missing alleles.}
}
\value{
A haplotype frequency table.
}
\description{
Estimate haplotype frequencies
}
