% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ld.R
\name{ld_by_group}
\alias{ld_by_group}
\title{Calculate LD separately by population group}
\usage{
ld_by_group(x, groups, ...)
}
\arguments{
\item{x}{An `ld_data` object.}

\item{groups}{Named group vector, group vector in sample order, or a two-column sample/group data frame.}

\item{...}{Arguments passed to `ld_compute`.}
}
\value{
A named list of `ld_result` objects with class `ld_grouped`.
}
\description{
Calculate LD separately by population group
}
