% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ld.R
\name{ld_compare}
\alias{ld_compare}
\title{Compare LD matrices between groups}
\usage{
ld_compare(grouped, metric = c("r2", "dprime"), reference = 1L)
}
\arguments{
\item{grouped}{Result from `ld_by_group`.}

\item{metric}{`r2` or `dprime`.}

\item{reference}{Name or index of the reference group.}
}
\value{
A list of group-minus-reference difference matrices.
}
\description{
Compare LD matrices between groups
}
