% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parse_region}
\alias{parse_region}
\title{Parse a genomic region}
\usage{
parse_region(region)
}
\arguments{
\item{region}{A string such as `chr1:1000-2000` or `chr1:1000:2000`.}
}
\value{
A one-row data frame with `chr`, `start`, and `end`.
}
\description{
Parse a genomic region
}
