% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ld.R
\name{plot_ld_decay}
\alias{plot_ld_decay}
\title{Plot an LD decay summary}
\usage{
plot_ld_decay(
  x,
  col = "#B2182B",
  ribbon_col = "#FDD0A2",
  xlab = "Physical distance (bp)",
  ylab = NULL,
  ...
)
}
\arguments{
\item{x}{Result from `ld_decay`.}

\item{col}{Line color.}

\item{ribbon_col}{Interquartile ribbon color.}

\item{xlab, ylab}{Axis labels.}

\item{...}{Additional graphical parameters.}
}
\value{
The input invisibly.
}
\description{
Plot an LD decay summary
}
