% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_plink}
\alias{read_plink}
\title{Read PLINK genotype files}
\usage{
read_plink(
  prefix,
  region = NULL,
  samples = NULL,
  min_maf = 0.05,
  max_missing = 0.25,
  max_het = 1,
  min_hwe_p = 0,
  quiet = FALSE
)
}
\arguments{
\item{prefix}{Prefix for BED/BIM/FAM or PED/MAP files.}

\item{region}{Optional region.}

\item{samples}{Optional sample vector or list file.}

\item{min_maf, max_missing, max_het, min_hwe_p}{Variant filters.}

\item{quiet}{Suppress filtering messages.}
}
\value{
An `ld_data` object.
}
\description{
Read PLINK genotype files
}
