% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{subset_ld_data}
\alias{subset_ld_data}
\title{Subset an LD genotype object}
\usage{
subset_ld_data(x, variants = NULL, samples = NULL)
}
\arguments{
\item{x}{An `ld_data` object.}

\item{variants}{Variant indices, IDs, or logical vector.}

\item{samples}{Sample indices, IDs, or logical vector.}
}
\value{
A new `ld_data` object with recalculated variant statistics.
}
\description{
Subset an LD genotype object
}
